Fix issue where database password isn't read because of the user
For example, when using the Passenger module in Apache, the user becomes
apache. However, the file containing the password isn't read because there are
no home directory for user apache. Thus, absolute path to the file is
hardcoded. You must change the location accordingly when using in a different
environment.