Add simple-captcha gem to prevent spam comments
Captcha on comment forms is added. The message is hard-coded and the user gets
redirected to the article page, thus, losing the comment he/she entered. This
needs to be reworked, but the current implementation should solve the problem
where some idiot posts tons of spam comments every now and then.